Add WASP-12_b analysis scripts and parameters
- Introduced `params_star.csv` containing stellar parameters for WASP-12_b.
- Added `run.py` script for analyzing transit timing variations (TTVs) and generating plots.
- Created `settings.csv` for configuration settings related to the analysis.
- Updated `allesfitter` to support pre-built transit models for TTV analysis, improving performance.
- Modified light curve evaluation to utilize a single global RoadRunnerModel instance, enhancing efficiency.
- Fixed MCMC sampling step calculation in `gp_decor.py` to ensure correct execution.

+++ allesfitter/exoworlds_rdx/lightcurves/gp_decor.py
@@ -468,7 +468,7 @@ def gp_decor(x,y,
                 
         #::: run evaluation
         logprint("\nRunning full MCMC")
-        sampler.run_mcmc(p0, total_steps/thin_by, thin_by=thin_by, progress=True);
+        sampler.run_mcmc(p0, int(total_steps/thin_by), thin_by=thin_by, progress=True);
     
     
     if multiprocess:
